+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 ---
 
+## [0.4.0](https://github.com/AnTonhoLAB/GGDevelopmentKit/releases/tag/0.4.0) - 2022-01-12
+### Added
+- GGPaginationSupport - Add pagination support to help collectionView/tableView pagination
+- UICollectionView - Reactive extension to know when the cell will be displayed and when the last cell will be displayed
+- SharedSequence - Reactive extension to unwrap optional values
+
 ## [0.3.2](https://github.com/AnTonhoLAB/GGDevelopmentKit/releases/tag/0.3.2) - 2022-01-07
 ### Fix
 - ErrorTracker - Turn ErrorTracker init public